Add season integration with selector and endpoints
The URL scheme for the main dashboard needs further thought. Currently,
players are the top level resource and other variables (currently only
season_id) are query parameters. E.g., `/players/2544?season_id=2019`.

The alternative would be to make seasons the top-level resource. In this
alternative, season_id would be a required parameter for the current player
dashboard. E.g., `/seasons/2019/players/2544`. This is easier for a human
to read. The drawback to this scheme is that it would limit future
feature velocity for features comparing across different seasons since a
new URL layout would be needed.

Perhaps this alternative URL could even be nested to future-proof it
against that drawback, e.g.`/playerdashboard/seasons/2019/players/2544`,
and the multi-season comparison dashboard could be nested under
`/comparison/`.