Remove conditional operators in JSX rendering null
In JSX/React, a conditional (ternary) operator is useful to
conditionally render components. If rendering either a component or
nothing, a conditional operator with an "else block" of `null` could be
used. React ignores `null` and `false` values when given them JSX
components. However, since `false` works too, the simpler way to do this
is to use the logical AND (`&&`) operator.