Change BarChart to use custom Voronoi implementation
This change removes the last usage of any major chart components from
the victory library. Before removing victory from the project
dependencies, the victorytheme needs to be refactored to a simple theme
and the ShootingSignature Legend needs to be either removed or
refactored (it is currently not being used at all).