Fix undefined hexagon bin color
This bug was due to a bin's shots being within maxDistance but the bin
center being beyond maxDistance. The lookup of league averages would
thus fail because it only fetches values from 0 up to maxDistance.

Now that this is being discussed though, it may make more sense to just
fetch league averages for all distances, which would also be more
mathematically correct. If the current setup causes another bug, this
alternative method should be implemented.